requirements.txt
numpy
pandas
openpyxl
pyarrow
ydata-profiling
sweetviz
matplotlib
seaborn
azure-storage-blob
azure-core
python-dotenv
jupyter
ipykernel
azure-ai-textanalytics
nltk
spacy
spacy-spanish-lemmatizer
supabase
psycopg2-binary
